requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.14.0
drf-spectacular>=0.28.0
djangorestframework-simplejwt>=5.2.2